Improve matching results
Instead of passing 'volume' & 'year' params to Metron's issue endpoint, let's return a larger result set.

That way we can compare the issue's cover hash to get the right one. Based on my initial testing this works pretty well, tho older issues with bad cover's may not give that great of results.

If it turns out this doesn't work for most user's, we can look at using the 'volume' and 'year' to narrow down the results.