Add --accept-only option to automatically accept single valid matches
As discussed in #180, if a search results in only a single valid match, even if the cover match doesn't meet the required threshold for an auto-match, if --accept-only is passed, then accept it anyway. This should result in a much smoother (less interactive) import for cases where the filename is already well-matched.